chore: remove editor references and move enrichEntity
